+++ src/main/java/com/tc/tech_challange/domain/endereco/EnderecoRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.web.bind.annotation.PathVariable;
 
-public interface EnderecoRepository extends JpaRepository<Endereco, Long> {
+public interface EnderecoRepository extends JpaRepository<Endereco, Integer> {
     Endereco getReferenceById(PathVariable id);
     Page<Endereco> findAll(Pageable paginacao);
 }
